+++ udata/core/reuse/api.py
@@ -99,7 +99,9 @@ class ReuseListAPI(API):
     @api.expect(Reuse.__index_parser__)
     @api.marshal_with(Reuse.__page_fields__)
     def get(self):
-        query = Reuse.objects.visible_by_user(current_user)
+        query = Reuse.objects.visible_by_user(
+            current_user, mongoengine.Q(private__ne=True, deleted=None)
+        )
         return Reuse.apply_sort_filters_and_pagination(query)
 
     @api.secure